requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==1.10.12
pydantic-settings==2.0.3
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
dateparser==1.1.8
openai==1.42.0
slowapi==0.1.9
python-dateutil==2.8.2
